Moves argless enum out
