+++ src/Type/index.js
@@ -1,5 +1,5 @@
 import { isList, isObject, matchPattern, listToObject, isConstructor } from '../common-utils';
-import ArgLessEnum from './ArgLessEnum';
+import ArgLessEnum from '../ArgLessEnum';
 
 export const all = (list, fn) => list.length === [...list].filter(fn).length;
 export const some = (list, fn) => !![...list].filter(fn).length;
